将ID生成策略转换为雪花算法

+++ evaluation/src/main/java/com/buguagaoshu/homework/evaluation/cache/ArticleTagCache.java
@@ -17,5 +17,5 @@ public class ArticleTagCache {
     private List<ArticleTagEntity> articleTagCaches;
 
 
-    private Map<Integer, ArticleTagEntity> integerArticleTagEntityMap;
+    private Map<Long, ArticleTagEntity> integerArticleTagEntityMap;
 }
